Add : Utility Function For Token
