+++ src/app/modules/auth/auth.controller.ts
@@ -4,19 +4,22 @@ import { sendResponse } from "../../utils/sendResponse";
 import httpStatus from "http-status-codes";
 import { authServices } from "./auth.service";
 import AppError from "../../errorHelpers/AppError";
+import { setAuthCookie } from "../../utils/setCookie";
 
 const credentialsLogin = catchAsync(async (req: Request, res: Response) => {
   const loginInfo = await authServices.credentialsLogin(req.body);
 
-  res.cookie("accessToken", loginInfo.accessToken, {
-    httpOnly: true,
-    secure: false,
-  });
+  // res.cookie("accessToken", loginInfo.accessToken, {
+  //   httpOnly: true,
+  //   secure: false,
+  // });
 
-  res.cookie("refreshToken", loginInfo.refreshToken, {
-    httpOnly: true,
-    secure: false,
-  });
+  // res.cookie("refreshToken", loginInfo.refreshToken, {
+  //   httpOnly: true,
+  //   secure: false,
+  // });
+
+  setAuthCookie(res, loginInfo);
 
   sendResponse(res, {
     success: true,
@@ -36,6 +39,13 @@ const getNewAccessToken = catchAsync(async (req: Request, res: Response) => {
   const tokenInfo = await authServices.getNewAccessToken(
     refreshToken as string
   );
+
+  // res.cookie("accessToken", tokenInfo.accessToken, {
+  //   httpOnly: true,
+  //   secure: false,
+  // });
+
+  setAuthCookie(res, tokenInfo);
 
   sendResponse(res, {
     success: true,
