+++ src/app/utils/userTokens.ts
@@ -1,6 +1,10 @@
+import { JwtPayload } from "jsonwebtoken";
 import { envVariables } from "../config/env";
-import { IUser } from "../modules/user/user.interface";
-import { generateToken } from "./jwt";
+import { IsActive, IUser } from "../modules/user/user.interface";
+import { generateToken, verifyToken } from "./jwt";
+import { User } from "../modules/user/user.model";
+import AppError from "../errorHelpers/AppError";
+import httpStatus from "http-status-codes";
 
 export const createUserTokens = (user: Partial<IUser>) => {
   const jwtPayload = {
@@ -22,4 +26,46 @@ export const createUserTokens = (user: Partial<IUser>) => {
   );
 
   return { accessToken, refreshToken };
+};
+
+export const createNewAccessTokenWithRefreshToken = async (
+  refreshToken: string
+) => {
+  const verifiedToken = verifyToken(
+    refreshToken,
+    envVariables.JWT_REFRESH_SECRET
+  ) as JwtPayload;
+
+  const isUserExist = await User.findOne({ email: verifiedToken.email });
+
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "User does not exist");
+  }
+
+  if (
+    isUserExist.isActive === IsActive.BLOCKED ||
+    isUserExist.isActive === IsActive.INACTIVE
+  ) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      `User is ${isUserExist.isActive}`
+    );
+  }
+
+  if (isUserExist.isDeleted) {
+    throw new AppError(httpStatus.BAD_REQUEST, "User is deleted");
+  }
+
+  const jwtPayload = {
+    userId: isUserExist._id,
+    email: isUserExist.email,
+    role: isUserExist.role,
+  };
+
+  const accessToken = generateToken(
+    jwtPayload,
+    envVariables.JWT_ACCESS_SECRET,
+    envVariables.JWT_ACCESS_EXPIRE
+  );
+  return accessToken;
 };
